+++ R/utils_web.R
@@ -20,6 +20,12 @@ get_crawl_delay <- function(crawl_delay = NULL,
                             use_delay = NULL) {
   rand_fast_delay <- runif(1, 0.5, 1.0)
 
+  # may have na
+  if (!is.null(use_delay)) {
+    if (is.na(use_delay)) use_delay <- NULL
+  }
+  
+  # return custom delay
   if (is.null(crawl_delay) && is.null(use_delay)) {
     return(rand_fast_delay)
   }